Finalise the rails test framework
